bug修复+DQN测试

+++ crypto_env/algorithm/moving_average.py
@@ -7,27 +7,27 @@ class MovingAverage(Algorithm):
     def __init__(self, price_feature_pos, short_window=30, long_window=180,
                  initial_cap=100):
         # simplified version of moving average. buy all cap / sell all cap at once.
-        self._wallet_usd = np.float64(initial_cap)
-        self._wallet_crypto = np.float64(0)
+        self._wallet_usd = np.float32(initial_cap)
+        self._wallet_crypto = np.float32(0)
         self._price_feature_pos = price_feature_pos
 
         # moving average things
-        self._short_window = deque([np.float64(0)], maxlen=short_window)
-        self._long_window = deque([np.float64(0)], maxlen=long_window)
+        self._short_window = deque([np.float32(0)], maxlen=short_window)
+        self._long_window = deque([np.float32(0)], maxlen=long_window)
         self._short_avg_history = [0]
         self._long_avg_history = [0]
 
     def _update_windows(self, price):
-        self._short_window.append(np.float64(price))
-        self._long_window.append(np.float64(price))
+        self._short_window.append(np.float32(price))
+        self._long_window.append(np.float32(price))
         self._short_avg_history.append(self._get_short_avg())
         self._long_avg_history.append(self._get_long_avg())
 
     def _get_long_avg(self):
-        return np.array(self._long_window, dtype=np.float64).mean()
+        return np.array(self._long_window, dtype=np.float32).mean()
 
     def _get_short_avg(self):
-        return np.array(self._short_window, dtype=np.float64).mean()
+        return np.array(self._short_window, dtype=np.float32).mean()
 
     def _generate_signal(self):
         cur_short = self._short_avg_history[-1]
@@ -43,7 +43,7 @@ class MovingAverage(Algorithm):
             return 2
 
     def take_action(self, observation, info=None):
-        price = np.float64(observation['features'][self._price_feature_pos])
+        price = np.float32(observation['features'][self._price_feature_pos])
         self._update_windows(price=price)
         signal = self._generate_signal()
         crypto_to_buy = self._wallet_usd / price
@@ -53,14 +53,14 @@ class MovingAverage(Algorithm):
         if signal == 0:  # buy
             action = dict(
                 signal=signal,
-                value=np.array([crypto_to_buy], dtype=np.float64)
+                value=np.array([crypto_to_buy], dtype=np.float32)
             )
             self._wallet_crypto += crypto_to_buy
             self._wallet_usd = 0
         if signal == 1:  # sell
             action = dict(
                 signal=signal,
-                value=np.array([crypto_to_sell], dtype=np.float64)
+                value=np.array([crypto_to_sell], dtype=np.float32)
             )
             self._wallet_usd += self._wallet_crypto * price
             self._wallet_crypto = 0
